tried updating the current database to include the music info but that is not updatting.
